Refactor construction hierarchy data
